Finalização das alterações
