+++ resources/views/Site/Vendas/index.blade.php
@@ -2,10 +2,6 @@
 
 @section('content')
 
-<link rel="stylesheet" href="{{asset('plugins/datatables-bs4/css/dataTables.bootstrap4.min.css')}}">
-<link rel="stylesheet" href="{{asset('plugins/datatables-responsive/css/responsive.bootstrap4.min.css')}}">
-<link rel="stylesheet" href="{{asset('plugins/datatables-buttons/css/buttons.bootstrap4.min.css')}}">
-
 <div class="content-header">
     <div class="container-fluid">
         <div class="row mb-2">
