+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\Api as RouteApi;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,3 +18,6 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
+
+Route::resource('saleitens', RouteApi\SaleItensController::class);
+Route::post('saleitens/listsaleitens/{sale_id}', [RouteApi\SaleItensController::class, 'list_sale_itens'])->name('get_list_itens_sale');
